Expand on data access across modules (#14)
If I understood the idea correctly accessing data from different modules can be done with data sources (of there is a data source that retrieves what is needed), but in most cases one module outputs data via an output, and that data gets wired in to another module via it's variables. I changed the text to reflect that.